backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
azure-identity==1.17.1
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-resource==23.1.0
azure-mgmt-compute==32.0.0
azure-mgmt-network==25.4.0
azure-mgmt-storage==21.1.0
azure-mgmt-sql==3.0.1
azure-mgmt-web==7.3.1
azure-mgmt-advisor==9.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-reservations==2.3.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-subscription==3.1.1
azure-mgmt-cognitiveservices>=13.5.0
httpx==0.27.2
pydantic==2.8.2
python-dateutil==2.9.0
anthropic==0.34.2
openai==1.51.0
